+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/common/request/OwnPageable.java
@@ -22,7 +22,9 @@ package org.eclipse.tractusx.traceability.common.request;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 import org.eclipse.tractusx.traceability.common.model.BaseRequestFieldMapper;
 import org.eclipse.tractusx.traceability.common.model.UnsupportedSearchCriteriaFieldException;
 import org.springframework.data.domain.PageRequest;
@@ -34,6 +36,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 @Data
+@Builder
+@NoArgsConstructor
 @AllArgsConstructor
 public class OwnPageable {
     private Integer page;