+++ CHANGELOG.md
@@ -15,6 +15,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 - Fixed security findings
+- Rework GET alerts and investigations endpoint to POST to send a request body
 - Fixed deadline overlap issues for Q-investigations in update menu
 - Fixed sorting of asPlanned parts
 - Fixed bug where applying a filter in one table erroneously affected sorting in the other table